Update downer to use the version from Nexuiz 1.1
With my buggy Intel 620 graphics card, there is an issue with the floor
texture "flickering" as I walk from the mortar to the jumppad.  By using
the older Nexuiz 1.1 map and BSP file, this flickering goes away.

Note that there is still a flicker on the floor in the basement walking
towards the mega-armor with the two cells with this buggy card/drivers,
but that flicker is more subtle.

There is also a flicker on the "toxic" map on the right hand wall when
walking up the stars to the mid-stair area with the medium health, but
it’s resonably subtle.