track001.ogg is from an old Quake3 map, not Nexuiz.  Attribute properly.
In more detail, this song comes from a map called quintdm3 made by
someone with the name “Brad”.  I have attempted to contact Brad,
and have, in the meantime, correctly attributed this song.

If I were to do it again, I would use a song from Nexuiz or Xonotic, but
I have been using this song since 2010, so it has been grandfathered in
to nexuiz-tiny.