Add texture which fixes issues with missing texture in Toxic map.
See https://github.com/samboy/nexuiz-tiny/issues/1

The inside ceiling above the electro in the Toxic map was missing.

I fixed this as follows:

cd data/textures/
mkdir e7/
cp e8/e8_base1.jpg e7/e7brnmetal.jpg

Yes, it was a different texture, but the point here is that it took
me a while to spot it, so we may as well save a bit of space by just
using another texture.