Put Toxic theme in its own .pk3 file to make it easier to remove.
When this is done, the Toxic map now uses the intro theme when played.
In addition, I have changed the Blood Prison map to use the Graphite theme
when the Blood Prison theme .pk3 is removed.  This way, when either or both
of these .pk3 files are removed, all maps still have themes.

Note that, when the Toxic theme is removed, there is no music that is
played after a match ends.

Removing both the Blood Prison and Toxic music .pk3 files allows nexuiz-tiny
to fit in 50 megabytes (just over 47.74 mebibytes).