downer.map: Update .map file to correspond to .bsp file
A while ago, I made a one-character change to the downer.bsp file so that
the skybox would not change the skybox in other Nexuiz maps.

I use a different skybox for Downer in this tiny version of Nexuiz:
The very beautiful "Heaven" skybox which did not exist at the time the
Downer map was made; since I included this skybox for the Graphite map,
I updated Downer to use the same skybox so that we would not have to
waste space by having another skybox in this compact version of Nexuiz.
Also, it looks better than the original Downer skybox.

The change was this (in Perl regular expression form):

s/sav-graysky1_skybox/sav-grayskyH_skybox/

To update the .map file I did the following two commands:

cat downer.map | perl -pe 's/sav-graysky1_skybox/sav-grayskyH_skybox/' > foo
mv foo downer.map